meson: Improve dependency search on MSVC builds
The upstream Visual Studio build files for the various dependencies
often do not generate pkg-config files for us, so we need to look for
these dependencies manually, if needed, before we use the fallbacks.

For FreeType, also support looking for FreeType using CMake, as the
CMake build configuration files use different naming and versioning
schemes for FreeType.