meson: Improve dependency search on MSVC builds
The upstream Visual Studio build files for the various dependencies
often do not generate pkg-config files for us, so we need to look for
these dependencies with an alternative approach like via CMake.

For FreeType, make sure we can find it via CMake as there are different
naming and versioning schemes for it.